.ipynb_checkpoints/requirements-checkpoint.txt
--extra-index-url https://download.pytorch.org/whl/cu128

numpy>=1.23.5,<2
typing-extensions>=4.8.0
opencv-python==4.10.0.84
cv2_enumerate_cameras==1.1.15
onnx==1.18.0
insightface==0.7.3
psutil==5.9.8
tk==0.1.0
customtkinter==5.2.2
pillow==11.1.0
torch; sys_platform != 'darwin'
torch==2.5.1; sys_platform == 'darwin'
torchvision; sys_platform != 'darwin'
torchvision==0.20.1; sys_platform == 'darwin'
onnxruntime-silicon==1.16.3; sys_platform == 'darwin' and platform_machine == 'arm64'
onnxruntime-gpu==1.22.0; sys_platform != 'darwin'
tensorflow; sys_platform != 'darwin'
opennsfw2==0.10.2
protobuf==4.25.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiortc==1.6.0
av==10.0.0
opencv-python==4.10.0.84
numpy>=1.23.5,<2
pillow==11.1.0
insightface==0.7.3
onnx==1.18.0
onnxruntime==1.22.0
typing-extensions>=4.8.0
psutil==5.9.8